+++ dessia_common/displays.py
@@ -60,6 +60,7 @@ class DisplaySetting:
 
 class DisplayObject:
     """ Container for data of display. A traceback can be set if display fails to be generated. """
+
     def __init__(self, type_: str, data: Union[JsonSerializable, List[JsonSerializable], str],
                  reference_path: str = '', traceback: str = '', name: str = ''):
         if type_ == "plot_data" and not is_sequence(data):